created role besed authorization
Created Admin, Manager, and Member roles and assigned the Admin and the Manager roles to two accounts. Every new user that regiters will be assigned a Member role. I restricted some of the pages to only users with the Admin or Manager role. I also made the Gallery vanigation link to appear only to users who are authorized to access it.